Move upload logic to serials service
