+++ SerialsManager/AppDelegate.swift
@@ -138,24 +138,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
                 self.loadingSheet.prepareForError(err.description)
         }
     }
-    
-    enum SerialsError: ErrorType, CustomStringConvertible {
-        case DownloadError(description: String)
-        case UploadError(description: String)
-        case Unauthorized
-        case NotValid
-        case Unknown
-        
-        var description: String {
-            switch self {
-                case .DownloadError(let description): return description
-                case .UploadError(let description): return description
-                case .Unauthorized: return "Unauthorized"
-                case .NotValid: return "JSON is not valid"
-                case .Unknown: return "Unknow error"
-            }
-        }
-    }
 
     func applicationWillTerminate(aNotification: NSNotification) {
         // Insert code here to tear down your application
@@ -207,141 +189,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
                 }
                 self.loadingSheet.prepareForError(err.description)
         }
-    }
-    
-    func uploadSerials(serials: [Serial]) -> Promise<[Serial]> {
-        let serialsWithSeasons =
-            serials.filter {
-                $0.seasons != nil
-        }
-        guard serialsWithSeasons.count != 0  else {
-            return Promise(serials)
-        }
-        return firstly { () -> Promise<[Files.FolderMetadata]> in
-                // create folders first because I don't want to write JSON file
-                // if an error happen
-                let promises: [Promise<Files.FolderMetadata>] =
-                    serialsWithSeasons.map { (serial: Serial) -> Promise<Files.FolderMetadata> in
-                        return self.createFolder("/" + serial.data.title)
-                }
-                return join(promises)
-            }.then { (_: [Files.FolderMetadata]) throws -> Promise<String> in
-                return Promise { resolve, reject in
-                    do {
-                        let serialsJSON = try self.toJSON(serials) { serial in
-                            let serialDict = NSMutableDictionary()
-                            serialDict.setValue(serial.data.title, forKey: "title")
-                            serialDict.setValue("/" + serial.data.title + "/seasons.json", forKey: "path")
-                            return serialDict
-                        }
-                        resolve(serialsJSON)
-                    } catch let error {
-                        reject(error)
-                    }
-                }
-            }.then { (serialsJSON: String) -> Promise<Files.FileMetadata> in
-                return self.uploadFile(
-                    "/serials.json",
-                    body: serialsJSON.dataUsingEncoding(NSUTF8StringEncoding)!
-                )
-            }.then { (_: Files.FileMetadata) -> [Serial] in
-                return serialsWithSeasons
-        }
-    }
-    
-    func uploadSeasons(seasons: [Season], prefix: String) -> Promise<[Season]> {
-        let seasonsWithChapters =
-            seasons.filter { $0.chapters != nil }
-        guard seasonsWithChapters.count != 0 else {
-            return Promise(seasons)
-        }
-        return firstly { () -> Promise<[Files.FolderMetadata]> in
-                // create folders first because I don't want to write JSON file
-                // if an error happen
-                let promises: [Promise<Files.FolderMetadata>] =
-                    seasonsWithChapters.map { (season: Season) -> Promise<Files.FolderMetadata> in
-                        let folder = prefix + "/" + season.data.title
-                        return self.createFolder(folder)
-                }
-                return join(promises)
-            }.then { (_: [Files.FolderMetadata]) throws -> Promise<String> in
-                return Promise { resolve, reject in
-                    do {
-                        let seasonsJSON = try self.toJSON(seasons) { season in
-                            let seasonDict = NSMutableDictionary()
-                            seasonDict.setValue(season.data.title, forKey: "title")
-                            seasonDict.setValue(prefix + "/" + season.data.title + "/chapters.json", forKey:    "path")
-                            return seasonDict
-                        }
-                        resolve(seasonsJSON)
-                    } catch let error {
-                        reject(error)
-                    }
-                }
-            }.then { (seasonsJSON: String) -> Promise<Files.FileMetadata> in
-                return self.uploadFile(
-                    prefix + "/seasons.json",
-                    body: seasonsJSON.dataUsingEncoding(NSUTF8StringEncoding)!
-                )
-            }.then { _ in
-                return seasonsWithChapters
-        }
-    }
-
-    func uploadChapters(chapters: [Chapter], prefix: String) -> Promise<[Chapter]> {
-        guard chapters.count != 0 else {
-            return Promise(chapters)
-        }
-        return firstly { () -> Promise<[Files.FileMetadata]> in
-                // uplaod files first because I don't want to write JSON file
-                // if an error happen
-                let promises =
-                    chapters.map { chapter in
-                        return self.uploadFile(
-                            prefix + "/" + chapter.data.title! + ".srt",
-                            body: chapter.data.raw!.dataUsingEncoding(NSUTF8StringEncoding)!
-                    )
-                }
-                return join(promises)
-            }.then { (_: [Files.FileMetadata]) throws -> Promise<String> in
-                return Promise { resolve, reject in
-                    do {
-                        let chaptersJSON = try self.toJSON(chapters) { chapter in
-                            let chapterDict = NSMutableDictionary()
-                            chapterDict.setValue(chapter.data.title, forKey: "title")
-                            let chapterSrtPath = prefix + "/" + chapter.data.title! + ".srt"
-                            chapterDict.setValue(chapterSrtPath, forKey: "path")
-                            return chapterDict
-                        }
-                        resolve(chaptersJSON)
-                    } catch let error {
-                        reject(error)
-                    }
-                }
-            }.then { (chaptersJSON: String) -> Promise<Files.FileMetadata> in
-                return self.uploadFile(
-                    prefix + "/chapters.json",
-                    body: chaptersJSON.dataUsingEncoding(NSUTF8StringEncoding)!
-                )
-            }.then { _ in
-                return chapters
-        }
-    }
-    
-    func toJSON<T>(items: [T], _ itemToDict: (T) -> NSMutableDictionary) throws -> String {
-        let toDicts: [NSMutableDictionary] = items.map(itemToDict)
-        let isValid = NSJSONSerialization.isValidJSONObject(toDicts)
-        if !isValid {
-            throw SerialsError.NotValid
-        }
-        var jsonString: String;
-        do {
-            let serialJSON = try NSJSONSerialization.dataWithJSONObject(toDicts, options: NSJSONWritingOptions())
-            jsonString = NSString(data: serialJSON, encoding: NSUTF8StringEncoding) as! String
-        } catch {
-            throw SerialsError.Unknown
-        }
-        return jsonString;
     }
 
 }
